+++ src/main/java/com/cburch/logisim/file/MouseMappings.java
@@ -62,7 +62,7 @@ public class MouseMappings {
         if (mods == cache_mods) {
             return cache_tool;
         } else {
-            Tool ret = map.get(Integer.valueOf(mods));
+            Tool ret = map.get(mods);
             cache_mods = mods;
             cache_tool = ret;
             return ret;
@@ -70,11 +70,11 @@ public class MouseMappings {
     }
 
     public Tool getToolFor(Integer mods) {
-        if (mods.intValue() == cache_mods) {
+        if (mods == cache_mods) {
             return cache_tool;
         } else {
             Tool ret = map.get(mods);
-            cache_mods = mods.intValue();
+            cache_mods = mods;
             cache_tool = ret;
             return ret;
         }
@@ -130,12 +130,12 @@ public class MouseMappings {
         }
 
         if (tool == null) {
-            Object old = map.remove(Integer.valueOf(mods));
+            Object old = map.remove(mods);
             if (old != null) {
                 fireMouseMappingsChanged();
             }
         } else {
-            Object old = map.put(Integer.valueOf(mods), tool);
+            Object old = map.put(mods, tool);
             if (old != tool) {
                 fireMouseMappingsChanged();
             }
@@ -143,7 +143,7 @@ public class MouseMappings {
     }
 
     public void setToolFor(Integer mods, Tool tool) {
-        if (mods.intValue() == cache_mods) {
+        if (mods == cache_mods) {
             cache_mods = -1;
         }
