+++ src/main/java/com/cburch/logisim/gui/opts/SimulateOptions.java
@@ -22,14 +22,14 @@ class SimulateOptions extends OptionsPanel {
     private MyListener myListener = new MyListener();
     private JLabel simLimitLabel = new JLabel();
     private JComboBox simLimit = new JComboBox(new Integer[]{
-            Integer.valueOf(200),
-            Integer.valueOf(500),
-            Integer.valueOf(1000),
-            Integer.valueOf(2000),
-            Integer.valueOf(5000),
-            Integer.valueOf(10000),
-            Integer.valueOf(20000),
-            Integer.valueOf(50000),
+            200,
+            500,
+            1000,
+            2000,
+            5000,
+            10000,
+            20000,
+            50000,
     });
     private JCheckBox simRandomness = new JCheckBox();
     private JLabel gateUndefinedLabel = new JLabel();
@@ -123,11 +123,11 @@ class SimulateOptions extends OptionsPanel {
         }
 
         private void loadSimLimit(Integer val) {
-            int value = val.intValue();
+            int value = val;
             ComboBoxModel model = simLimit.getModel();
             for (int i = 0; i < model.getSize(); i++) {
                 Integer opt = (Integer) model.getElementAt(i);
-                if (opt.intValue() == value) {
+                if (opt == value) {
                     simLimit.setSelectedItem(opt);
                 }
             }
@@ -138,7 +138,7 @@ class SimulateOptions extends OptionsPanel {
         }
 
         private void loadSimRandomness(Integer val) {
-            simRandomness.setSelected(val.intValue() > 0);
+            simRandomness.setSelected(val > 0);
         }
     }
 }
