+++ src/main/java/com/cburch/logisim/std/memory/CounterAttributes.java
@@ -21,7 +21,7 @@ class CounterAttributes extends AbstractAttributeSet {
                 StdAttr.EDGE_TRIGGER,
                 StdAttr.LABEL, StdAttr.LABEL_FONT
         }, new Object[]{
-                BitWidth.create(8), Integer.valueOf(0xFF),
+                BitWidth.create(8), 0xFF,
                 Counter.ON_GOAL_WRAP,
                 StdAttr.TRIG_RISING,
                 "", StdAttr.DEFAULT_LABEL_FONT
@@ -57,23 +57,23 @@ class CounterAttributes extends AbstractAttributeSet {
             int oldW = oldWidth.getWidth();
             int newW = newWidth.getWidth();
             Integer oldValObj = base.getValue(Counter.ATTR_MAX);
-            int oldVal = oldValObj.intValue();
+            int oldVal = oldValObj;
             base.setValue(StdAttr.WIDTH, newWidth);
             if (newW > oldW) {
-                newMax = Integer.valueOf(newWidth.getMask());
+                newMax = newWidth.getMask();
             } else {
                 int v = oldVal & newWidth.getMask();
                 if (v != oldVal) {
-                    Integer newValObj = Integer.valueOf(v);
+                    Integer newValObj = v;
                     base.setValue(Counter.ATTR_MAX, newValObj);
                     fireAttributeValueChanged(Counter.ATTR_MAX, newValObj);
                 }
             }
             fireAttributeValueChanged(StdAttr.WIDTH, newWidth);
         } else if (attr == Counter.ATTR_MAX) {
-            int oldVal = base.getValue(Counter.ATTR_MAX).intValue();
+            int oldVal = base.getValue(Counter.ATTR_MAX);
             BitWidth width = base.getValue(StdAttr.WIDTH);
-            int newVal = ((Integer) value).intValue() & width.getMask();
+            int newVal = (Integer) value & width.getMask();
             if (newVal != oldVal) {
                 @SuppressWarnings("unchecked")
                 V val = (V) Integer.valueOf(newVal);
